+++ examples/warandpeace/countwords.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cat - | gunzip | grep -o -E "(\\w|')+" | grep -v -P '^\d' | tr '[:upper:]' '[:lower:]' | sort | uniq  | wc -l