+++ mod.ts
@@ -2,6 +2,7 @@ export * from "./runner.ts";
 export * from "./runners/proc-group.ts";
 export * from "./runners/handlers/string.ts";
 export * from "./runners/handlers/bytes.ts";
+export * from "./runners/handlers/reader.ts";
 export * from "./runners/handlers/string-iterable.ts";
 export * from "./runners/handlers/bytes-iterable.ts";
 export * from "./runners/handlers/empty.ts";