+++ runners/handlers/bytes-iterable.ts
@@ -49,7 +49,7 @@ export class BytesIterableInputHandler
       }
       await bw.flush();
     } catch (e) {
-      if (e instanceof Deno.errors.BrokenPipe) {
+      if (e instanceof Deno.errors.BrokenPipe || e instanceof Deno.errors.Interrupted) {
         // Ignore.
       } else {
         throw e;