+++ runners/handlers/string-iterable.ts
@@ -56,7 +56,7 @@ export class StringIterableInputHandler
       }
       await bw.flush();
     } catch (e) {
-      if (e instanceof Deno.errors.BrokenPipe) {
+      if (e instanceof Deno.errors.BrokenPipe || e instanceof Deno.errors.Interrupted) {
         // Ignore.
       } else {
         throw e;