+++ release-notes.md
@@ -2,8 +2,8 @@
 
 ## 0.7.0 _Prerelease_
 
- - **fix** Added defensive close to all process resources on process exit in output handlers.
-
+- **fix** Added defensive close to all process resources on process exit in
+  output handlers.
 
 ## 0.6.1 _Prerelease_